Better support hard-source
- Improve cache name for hard-source to distinguish the compilations
  with
- Add transform dependency when a Module is returned to the
  EntryReferencePlugin

When hard-source or other plugins return a Module during the resolver
step of the normal module factory, the EntryReferencePlugin should make
sure there is an entry for that file. As well it will add a dependency
to the entry to ensure its included.

For self-referencing modules instead of returning a reference module
return the module directly. With hard-source for self-referencing
modules the reference module will not be created. Since the
self-referencing module is in the chunk a require directly to it is
fine.